Adjust classes based on inspections, adjust detekt inspections

+++ src/main/kotlin/com/edd/jelly/util/Units.kt
@@ -17,7 +17,7 @@ object Units {
 }
 
 private val ONE_RADIAN = (180 / Math.PI).toFloat()
-private val ONE_DEGREE = (1 / ONE_RADIAN).toFloat()
+private val ONE_DEGREE = (1 / ONE_RADIAN)
 
 /**
  * Get number value as meters.
